+++ src/game/tests.rs
@@ -377,6 +377,7 @@ mod vintage_cube_figure_of_destiny;
 mod vintage_cube_flash;
 mod vintage_cube_flickerwisp;
 mod vintage_cube_force_of_negation;
+mod vintage_cube_forth_eorlingas;
 mod vintage_cube_gadgeteer;
 mod vintage_cube_galvanic_blast;
 mod vintage_cube_gau;